requirements.txt
aniso8601==7.0.0
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
django-graphiql==0.4.4
get==2019.4.13
graphene==2.1.8
graphene-django==2.5.0
graphql-core==2.2.1
graphql-relay==2.0.0
idna==2.8
post==2019.4.13
promise==2.2.1
public==2019.4.13
pycodestyle==2.5.0
pytz==2019.2
query-string==2019.4.13
requests==2.22.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
